Fixed unsupported inline asm call.
This commit fixes type mismatches in an inline asm call. Calls of this
type are not compatabile with the Clang front-end compiler for LLVM.
This fix allows support for both GCC and Clang compilers.

+++ c/ldv-linux-3.4-simple/43_1a_cilled_true-unreach-call_ok_nondet_linux-43_1a-drivers--mtd--devices--doc2001plus.ko-ldv_main0_sequence_infinite_withcheck_stateful.cil.out.c
@@ -2538,7 +2538,7 @@ void ldv_spin_unlock(void) ;
 int ldv_spin_trylock(void) ;
 #line 397 "/home/zakharov/launch/inst/current/envs/linux-3.4/linux-3.4/arch/x86/include/asm/bitops.h"
 __inline static int ffs(int x ) 
-{ int r ;
+{ long r ;
   long tmp ;
 
   {
